Solution for Etch-A-Sketch problem from The Odin Project
Create a container to hold divs created by a Javascript loop
Ask user to enter in how many cells they want to generate, no more than 100x100. Grid on screen will be square and equal to the number of input cells by the number of input cells.
Assign each div created by loop an event listener that listens for a hover event and then changes the background color of the div to black once passed over.